+++ packages/.vitepress/shared/unocss.theme.ts
@@ -1052,11 +1052,14 @@ export const fontWeight = {
   black: "900"
 };
 export const breakpoints = {
-  sm: "640px",
+  sm: "576px",
   md: "768px",
   lg: "1024px",
   xl: "1280px",
-  "2xl": "1536px"
+  "2xl": "1536px",
+  xs: "480px",
+  "996px": "996px",
+  "1140px": "1140px"
 };
 export const verticalBreakpoints = {
   sm: "640px",
@@ -1236,7 +1239,8 @@ export const duration = {
   "700": "700ms",
   "1000": "1000ms",
   DEFAULT: "150ms",
-  none: "0s"
+  none: "0s",
+  colors: "300ms"
 };
 export const ringWidth = {
   DEFAULT: "3px",