+++ packages/snippets/src/ant-design-x/index.ts
@@ -5,6 +5,7 @@ export {
   XStream,
 } from '@ant-design/x';
 export * from '@mixte/snippets/ant-design-x/bubble';
+export * from '@mixte/snippets/ant-design-x/conversations';
 export * from '@mixte/snippets/ant-design-x/sender';
 export {
   defineBubbleListRoles,